# Conflicts:
#	FrontEnd/app/src/main/java/cat/copernic/grup4/gamedex/videogames/ui/screens/ViewGames.kt
#	FrontEnd/app/src/main/res/values/strings.xml

+++ BackEnd/gamedex/src/main/java/cat/copernic/gamedex/logic/CategoryLogic.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import cat.copernic.gamedex.entity.Category;
 import cat.copernic.gamedex.repository.CategoryRepository;
+import cat.copernic.gamedex.repository.VideogameRepository;
 
 /**
  * Lògica de negoci per gestionar les categories.
@@ -17,6 +18,9 @@ public class CategoryLogic {
 
     @Autowired
     private CategoryRepository categoryRepository;
+
+    @Autowired
+    private VideogameRepository videogameRepository;
 
     /**
      * Crea una nova categoria.
@@ -40,6 +44,7 @@ public class CategoryLogic {
 
     /**
      * Elimina una categoria pel seu nom.
+     * I si la categoria està en ús, llavors no es pot eliminar.
      *
      * @param nameCategory El nom de la categoria a eliminar.
      */
@@ -48,6 +53,9 @@ public class CategoryLogic {
             Optional<Category> category = categoryRepository.findById(nameCategory);
             if (category.isEmpty()) {
                 throw new RuntimeException("Category not found");
+            }
+            if (!videogameRepository.findByCategory(nameCategory).isEmpty()) {
+                throw new RuntimeException("Category is in use");
             }
             categoryRepository.deleteById(nameCategory);
         } catch (RuntimeException e) {
